requirements.txt
Flask==2.0.3
pandas==1.3.5
fpdf2==2.5.7
openpyxl==3.0.10
xlrd==2.0.1